Move edge cognition out of memory package
