+++ tests/unit/test_chat_websearch_net.py
@@ -84,7 +84,7 @@ def chat_env(monkeypatch):
         def metacognitive_context(self, *a, **kw):
             return ""
 
-    from cognition.memory import edge_state
+    from cognition import edge_state
     monkeypatch.setattr(edge_state, "for_identity", lambda uid: EdgeState())
     monkeypatch.setattr(think_module.bioclock, "current_datetime_block", lambda: "")
     monkeypatch.setattr(think, "_resolve_mem_kb", lambda *a, **kw: ([], ""))
